HW-3 불필요한 Interface 제거
