+++ wooooooak/StudyApp/app/src/main/java/wooooooak/com/studyapp/ui/base/ItemContract.kt
@@ -1,7 +1,17 @@
 package wooooooak.com.studyapp.ui.base
 
 interface ItemContract {
-    interface View<T> : ItemView<T>
+    interface View<T> {
+        fun renderItems(items: List<T>)
 
-    interface Presenter<T> : ItemPresenter<T>
+        fun renderWebView(url: String)
+
+        fun renderErrorToast(message: String?)
+    }
+
+    interface Presenter<T> {
+        suspend fun fetchItemsWithNewTitle(title: String)
+
+        suspend fun fetchMoreItems(originList: List<T>, index: Int = 0)
+    }
 }
