--- wooooooak/StudyApp/app/src/main/java/wooooooak/com/studyapp/ui/base/ItemPresenter.kt
@@ -1,7 +1,0 @@
-package wooooooak.com.studyapp.ui.base
-
-interface ItemPresenter<T> {
-    suspend fun fetchItemsWithNewTitle(title: String)
-
-    suspend fun fetchMoreItems(originList: List<T>, index: Int = 0)
-}
