--- wooooooak/StudyApp/app/src/main/java/wooooooak/com/studyapp/ui/base/ItemView.kt
@@ -1,9 +1,0 @@
-package wooooooak.com.studyapp.ui.base
-
-interface ItemView<T> {
-    fun renderItems(items: List<T>)
-
-    fun renderWebView(url: String)
-
-    fun renderErrorToast(message: String?)
-}
